requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
streamlit>=1.25.0
python-dotenv>=1.0.0